Rename legacy declarative compatibility module
