+++ declarative/tests/compatibility/legacy-4-module/src/main/java/io/helidon/declarative/tests/compatibility/v4/LegacyScheduledTask.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.declarative.tests.compatibility.v44;
+package io.helidon.declarative.tests.compatibility.v4;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicInteger;
